fix(scripts): use dtype= instead of deprecated torch_dtype= for transformers 5.x
transformers 5.x deprecated the torch_dtype kwarg on from_pretrained in
favor of dtype. The old kwarg still works but emits a deprecation warning;
switching to dtype= keeps model loading clean under the >=5.12.1 bump (#68).

Verified on transformers 5.9.0: dtype= loads with zero deprecation warnings.

+++ scripts/_runtime_common.py
@@ -19,7 +19,7 @@ def load_model_and_tokenizer(
 ) -> tuple[Any, Any]:
     model = cast(
         Any,
-        AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=dtype),
+        AutoModelForCausalLM.from_pretrained(model_name, dtype=dtype),
     )
     model.eval()
     tokenizer = cast(Any, AutoTokenizer.from_pretrained(model_name))
